+++ src/ui-kit/components/NetworkImage/NetworkImage.tsx
@@ -11,6 +11,9 @@ import OptimismImage from "../../images/optimism.png";
 import BobaImage from "../../images/boba.png";
 import SkaleImage from "../../images/skale.png";
 import CronosImage from "../../images/cronos.png";
+import CeloImage from "../../images/celo.jpg";
+import OKXChainImage from "../../images/okxchain.png";
+import GnosisImage from "../../images/gnosis.jpg";
 
 export type NetworkType =
     | "eth"
@@ -23,6 +26,9 @@ export type NetworkType =
     | "boba"
     | "skale"
     | "cronos"
+    | "celo"
+    | "okxchain"
+    | "gnosis"
     | string;
 
 interface NetworkImageProps {
@@ -57,6 +63,12 @@ const getNetworkImage = (network?: NetworkType) => {
             return SkaleImage;
         case "cronos":
             return CronosImage;
+        case "celo":
+            return CeloImage;
+        case "okxchain":
+            return OKXChainImage;
+        case "gnosis":
+            return GnosisImage;
         default:
             return DefaultImage;
     }